Checking valid authtoken but I don't know the right sqlite syntax/method to subtract datetime. CHANGED TO JAVA SECONDS
Load still needs to check for invalid or missing elements in the arrays.

For some reason,

success = (success && database.getUserDao().addUser(user))

doesn't work and does REALLY weird things like call the authToken service.

FOR SOME OTHER WEIRD REASON
the User.findUser() method is pulling the primary key from the user table when I look up #1, or at least it's storing it in the wrong part of the User. I don' t know if the Event and Person lookups need to watch for that too.