SettingsActivity implements nothing and doesn't use a fragment. Layout seems to look decent.
Map menu is able to check for the activity (mainactivity or mapactivity) and change accoringly.
The other activities will need the up (go to top) button, but preserve settings if possible.

Toolbar/Actionbar action buttons visible. Now need to make them clickable and go to the respective fragments.

Info panel on map Fragment appears. Need to fill it with the right stuff and make the icon dynamic