SettingsActivity wired up except for toggles.
In the middle of making the Settings Spinners work. Making a Settings class in Model.

SettingsActivity Re-Sync data works.

Menu buttons open their respective activities, PersonActivity and MapActivity bare bones. Need to make layouts for all but SettingsActivity

SettingsActivity layout is perfect (for my phone!)
SettingsActivity implements nothing and doesn't use a fragment.

Map menu is able to check for the activity (mainactivity or mapactivity) and change accoringly.
The other activities will need the up (go to top) button, but preserve settings if possible.

Toolbar/Actionbar action buttons visible. Now need to make them clickable and go to the respective fragments.

Info panel on map Fragment appears. Need to fill it with the right stuff and make the icon dynamic