Oooh I can click on events from the PersonActivity now :) Need to change the menu to blank except the back button.
Family lines ARE working even thought I thought they weren't
Spouse line draws to appropriate earliest event in spouse life

Life story Lines are drawing with changes in thickness
Filters confirmed working but only programmed.
Filters are programmed in but FilterActivity has no layout nor is it wired up.
Fixed issue with the user's Person not having the same personID in the User table of the database and the Person table of the database.

SettingsActivity fully wired up including toggles. Now the map needs to draw lines according to the settings.
I might work on the FilterActivity next.

In the middle of making the Settings Spinners work. Making a Settings class in Model.

SettingsActivity Re-Sync data works.

Menu buttons open their respective activities, PersonActivity and MapActivity bare bones. Need to make layouts for all but SettingsActivity

SettingsActivity layout is perfect (for my phone!)
SettingsActivity implements nothing and doesn't use a fragment.

Map menu is able to check for the activity (mainactivity or mapactivity) and change accoringly.
The other activities will need the up (go to top) button, but preserve settings if possible.

Toolbar/Actionbar action buttons visible. Now need to make them clickable and go to the respective fragments.

Info panel on map Fragment appears. Need to fill it with the right stuff and make the icon dynamic